pyproject.toml
[project]
name = "personal-rag-chatbot"
version = "0.1.0"
description = "RAG-Chatbot of Mr.Poornachandra A N for Professional use."
authors = [
    {name = "Poornachandra.A.N",email = "poornachandra308@gmail.com"}
]

readme = "README.md"

[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.11,<4.0"
faiss-cpu = "^1.7.4"
python-dotenv = "^1.0.0"
openai = "^1.12.0"
streamlit = "^1.31.0"
PyMuPDF = "^1.23.0"
pdfplumber = "^0.10.0"
sentence-transformers = "^2.3.0"
rich = "^13.7.0"
fastapi = "^0.109.0"
uvicorn = "^0.27.0"
tiktoken = "^0.5.2"
hf-xet = "^1.1.7"
langchain = "^0.3.27"
numpy = "^2.3.2"
langchain-community = "^0.3.27"
langchain-google-genai = "^2.1.9"
beautifulsoup4 = "^4.13.4"
docx2txt = "^0.9"
python-docx = "^1.2.0"
easyocr = "^1.7.2"
langdetect = "^1.0.9"
pytesseract = "^0.3.13"
googletrans = "^4.0.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
black = "^24.0.0"
isort = "^5.13.0"
flake8 = "^7.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
rag-cli = "src.interface:cli_main"
rag-server = "src.interface:server_main"